% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{adjust_timestamps}
\alias{adjust_timestamps}
\title{Adjust the timestamped data
 When the time is reset in mpc (e.g, if we restart), time intervals are messed up.
 This fixes them by assuming zero delay after the restart.}
\usage{
adjust_timestamps(time)
}
\arguments{
\item{time}{a vector of timestamps}
}
\value{
adjusted timestamps
}
\description{
Adjust the timestamped data
 When the time is reset in mpc (e.g, if we restart), time intervals are messed up.
 This fixes them by assuming zero delay after the restart.
}
