% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventcodes.R
\name{codefor}
\alias{codefor}
\title{Return an event code for a single event
 preserves negation (i.e., "-trial_start)}
\usage{
codefor(events, codestr)
}
\arguments{
\item{events}{Events in TSData, must be a factor}

\item{codestr}{a single event}
}
\value{
the event code associated with the event
}
\description{
Return an event code for a single event
 preserves negation (i.e., "-trial_start)
}
\examples{
eventcodes = codefor(data, "trial_start")
eventcodes = codefor(data, "-trial_start")
}
