% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventcodes.R
\name{codesfor}
\alias{codesfor}
\title{Return event codes for a list of events}
\usage{
codesfor(events, codes)
}
\arguments{
\item{events}{events in TSData, must be a factor}

\item{codes}{list of event names}
}
\value{
event codes associated with the events
}
\description{
Return event codes for a list of events
}
\examples{
eventcodes = codesfor(data, c("trial_start", "trial_end"))
}
