% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventcodes.R
\name{convert_codes_to_events}
\alias{convert_codes_to_events}
\title{Convert numeric event codes to named event codes}
\usage{
convert_codes_to_events(codes, eventcodes)
}
\arguments{
\item{codes}{List of numeric event codes}

\item{eventcodes}{List of named:numeric event code mappings}
}
\value{
list of named event codes for each of the numeric codes in 'codes'
}
\description{
Convert numeric event codes to named event codes
}
