% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{diagnostics}
\alias{diagnostics}
\title{Return diagnostics for common MedPC errors}
\usage{
diagnostics(events)
}
\arguments{
\item{events}{event codes}
}
\value{
an data_frame for every on event type
}
\description{
Return diagnostics for common MedPC errors
}
\examples{
out = diagnostics(events)
}
