% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{feeding_periods}
\alias{feeding_periods}
\title{Return feeding periods}
\usage{
feeding_periods(events)
}
\arguments{
\item{events}{event codes}
}
\value{
an array of feeding period numbers for each event
}
\description{
Return feeding periods
}
\examples{
periods = feeding_periods(events)
}
