% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{find_files}
\alias{find_files}
\title{Find all files in a path}
\usage{
find_files(FilePath)
}
\arguments{
\item{FilePath}{location and pattern to look for}
}
\value{
all files matching a pattern
}
\description{
Find all files in a path
}
\examples{
 find_files("~/Dropbox/Data/h/*")
 find_files("~Dropbox/Data/h/*.csv")
 find_files("~Dropbox/Data/h/*.999")
}
