% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitest.R
\name{habitest_load_file}
\alias{habitest_load_file}
\title{Load a single habitest file}
\usage{
habitest_load_file(file)
}
\arguments{
\item{file}{file to load}
}
\value{
data.frame of a habitest file
}
\description{
Load a single habitest file
}
\examples{
data = habitest_load_file(file)
}
