% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitest.R
\name{habitest_load_files}
\alias{habitest_load_files}
\title{Load all habitest files}
\usage{
habitest_load_files(files)
}
\arguments{
\item{files}{files to load}
}
\value{
data.frame with all files together
}
\description{
Load all habitest files
}
\examples{
data = find_files_bytype("csv", "./data/") \%>\%
       habitest_load_files()
}
