% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc_load_file}
\alias{mpc_load_file}
\title{Load a single mpc file}
\usage{
mpc_load_file(file, resolution = 0.01)
}
\arguments{
\item{file}{file to load}

\item{resolution}{the time resolution}
}
\value{
data.frame of a habitest file
}
\description{
Load a single mpc file
}
\examples{
data = mpc_load_file(file)
}
