% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc_load_files}
\alias{mpc_load_files}
\title{Load all mpc files}
\usage{
mpc_load_files(files, resolution = 0.01)
}
\arguments{
\item{files}{files to load}

\item{resolution}{the time resolution}
}
\value{
data.frame with all files together
}
\description{
Load all mpc files
}
\examples{
data = find_files("/Users/dfreesto/Dropbox/Data/h/*.999") \%>\%
       mpc_load_files()
}
