% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc.R
\name{mpc_tidy}
\alias{mpc_tidy}
\title{Cleanup mpc data.}
\usage{
mpc_tidy(df, eventcodes = NULL, files = NULL)
}
\arguments{
\item{df}{data.frame of loaded mpc files}

\item{eventcodes}{A data_frame of event codes}

\item{files}{A list of csv files that specify event codes}
}
\value{
data.frame with named event codes, and session numbers
}
\description{
Cleanup mpc data.
}
\examples{
data = find_files("~/Dropbox/Data/h/*.999") \%>\%
       mpc_load_files() \%>\%
       mpc_tidy(eventCodeFiles)
}
