% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster}
\alias{raster}
\title{Return raster information for each trial}
\usage{
raster(event, pattern)
}
\arguments{
\item{event}{named event column from a data frame}

\item{pattern}{the c(on, off, end) event information}
}
\value{
data.frame(trial, ontime, offtime) for each state event
}
\description{
Return raster information for each trial
}
\examples{
raster(data, c("LeftLeverOn", "LeftLeverOff", "trial_end"))
}
