% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{rasterplot}
\alias{rasterplot}
\title{Add a raster plot as a geom_segment to an existing ggplot}
\usage{
rasterplot(data, pattern, ax, color = "blue")
}
\arguments{
\item{data}{the data.frame}

\item{pattern}{the c(on, off, end) event information}

\item{ax}{The ggplot handle}

\item{color}{the color of the points}
}
\value{
ax the updated ggplot handle
}
\description{
Add a raster plot as a geom_segment to an existing ggplot
}
\examples{
raster(data, c("LeftLeverOn", "LeftLeverOff", "trial_end"))
}
