% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventcodes.R
\name{read_eventcode}
\alias{read_eventcode}
\title{Load a single file of event codes}
\usage{
read_eventcode(File)
}
\arguments{
\item{File}{event code file to load}
}
\value{
data.frame of event codes
}
\description{
Load a single file of event codes
}
