% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventcodes.R
\name{read_eventcodes}
\alias{read_eventcodes}
\title{Load several event code files into a data.frame}
\usage{
read_eventcodes(Files)
}
\arguments{
\item{Files}{event code files to load}
}
\value{
data.frame will loaded event codes
}
\description{
Load several event code files into a data.frame
}
