% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{session_fromdate}
\alias{session_fromdate}
\title{Session information from dates}
\usage{
session_fromdate(date)
}
\arguments{
\item{date}{dates for each session}
}
\value{
session numbers
}
\description{
Session information from dates
}
\examples{
data = habitest_load_file(file) \%>\%
       mutate(session=habitest_session(date))
}
