% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{time_of_day}
\alias{time_of_day}
\title{Return the time of day given a known anchor time}
\usage{
time_of_day(date, time, event, anchor_time, anchor_event)
}
\arguments{
\item{date}{date vector}

\item{time}{time vector}

\item{event}{event codes}

\item{anchor_time}{the known wall time of the anchor event}

\item{anchor_event}{the event for which the wall time is known}
}
\value{
an array of datetimes relative to the anchor event
}
\description{
Return the time of day given a known anchor time
}
\examples{
timeofday = time_of_day(date, time, event, anchor_time, anchor_event)
}
