% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{trialdef}
\alias{trialdef}
\title{Return trial numbers}
\usage{
trialdef(events, pattern, fromfirst = FALSE)
}
\arguments{
\item{events}{event codes}

\item{pattern}{a list of event codes for pattern search}
}
\value{
an array of trial numbers for each event
}
\description{
Return trial numbers
}
\examples{
trial = trialdef(events, pattern)
}
